# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Etscript API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /payments/checkout:
    post:
      operationId: createCheckout
      tags: [payments]
      summary: Initialize a payment and get a checkout URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentCheckoutInput"
      responses:
        "200":
          description: Checkout initialized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentCheckout"
        "400":
          description: Invalid request
        "401":
          description: Unauthorized
  /payments/verify:
    get:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a payment by reference after checkout
      parameters:
        - name: reference
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerification"
        "404":
          description: Transaction not found
  /payments/export-access/{jobId}:
    get:
      operationId: getExportAccess
      tags: [payments]
      summary: Whether the user can download clean (non-watermarked) exports for a job
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Export access status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportAccess"
...
components:
  schemas:
...
    PaymentCheckoutInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [payg_export, premium_subscription, premium_quarterly, premium_annual, lifetime_access]
        jobId:
          type: ["integer", "null"]
    PaymentCheckout:
      type: object
      required: [authorizationUrl, reference]
      properties:
        authorizationUrl:
          type: string
        reference:
          type: string
    PaymentVerification:
      type: object
      required: [status, type]
      properties:
        status:
          type: string
          enum: [success, pending, failed]
        type:
          type: string
          enum: [payg_export, premium_subscription, premium_quarterly, premium_annual, lifetime_access]
        jobId:
          type: ["integer", "null"]
    ExportAccess:
      type: object
      required: [canDownloadClean, plan, jobPaid]
      properties:
        canDownloadClean:
          type: boolean
        plan:
          type: string
          enum: [free, premium]
        jobPaid:
          type: boolean
...